requirements.txt
fastapi
nicegui
uvicorn
python-dotenv
requests
tqdm
googlemaps
docxtpl
pandas
tenacity
